requirements_docker.txt
# Application Configuration Dependencies
python-dotenv==0.15.0

# Service Configuration Dependencies
requests==2.25.0

# Application Functional Dependencies
numpy==1.19.3
scipy==1.5.4
matplotlib==3.3.3
holidays==0.10.4

# Server Dependencies
django==3.1.6
gunicorn==20.0.4
whitenoise==5.1.0
django-cors-headers==3.7.0
django-referrer-policy==1.0.0